抽象化のレベルあげた
